+++ examples/power_manager/go.mod
@@ -1,0 +1,3 @@
+module github.com/vhive-serverless/vhive/examples/power_manager
+
+go 1.19
