--- power_manager/go.mod
@@ -1,5 +1,0 @@
-module github.com/vhive-serverless/vhive/examples/deployer
-
-go 1.19
-
-replace github.com/vhive-serverless/vhive/examples/deployer => ../deployer
