--- power_manager/workload_sensitivity_exp.go
@@ -1,88 +1,0 @@
-package main
-
-import (
-	"encoding/csv"
-	"fmt"
-	"os"
-	"os/exec"
-	"strconv"
-	"time"
-)
-
-func setPowerProfileToNode(freq int64) error {
-	// powerConfig
-	command := fmt.Sprintf("kubectl apply -f - <<EOF\napiVersion: \"power.intel.com/v1\"\nkind: PowerConfig\nmetadata:\n  name: power-config\n  namespace: intel-power\nspec:\n powerNodeSelector:\n    kubernetes.io/hostname: node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us\n powerProfiles:\n    - \"performance\"\nEOF")
-	cmd := exec.Command("bash", "-c", command)
-	_, err := cmd.CombinedOutput()
-	if err != nil {
-		return err
-	}
-
-	// performanceProfile w freq
-	command = fmt.Sprintf("kubectl apply -f - <<EOF\napiVersion: \"power.intel.com/v1\"\nkind: PowerProfile\nmetadata:\n  name: performance\n  namespace: intel-power\nspec:\n  name: \"performance\"\n  max: %d\n  min: %d\n  shared: true\n  governor: \"performance\"\nEOF", freq, freq)
-	cmd = exec.Command("bash", "-c", command)
-
-	_, err = cmd.CombinedOutput()
-	if err != nil {
-		return err
-	}
-
-	// apply to node
-	command = fmt.Sprintf("kubectl apply -f - <<EOF\napiVersion: \"power.intel.com/v1\"\nkind: PowerWorkload\nmetadata:\n  name: performance-node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us-workload\n  namespace: intel-power\nspec:\n  name: \"performance-node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us-workload\"\n  allCores: true\n  powerNodeSelector:\n    kubernetes.io/hostname: node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us\n  powerProfile: \"performance\"\nEOF")
-	cmd = exec.Command("bash", "-c", command)
-
-	_, err = cmd.CombinedOutput()
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func main() {
-	file, err := os.Create("metrics.csv")
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-
-	writer := csv.NewWriter(file)
-	defer writer.Flush()
-
-	err = writer.Write(append([]string{"startTime", "endTime", "latency"}))
-	if err != nil {
-		fmt.Printf("Error writing metrics to the CSV file: %v\n", err)
-	}
-
-	frequencies := []int64{1200, 1400, 1600, 1800, 2000, 2200, 2400, 2600}
-	for i := 0; i < len(frequencies); i++ {
-		err := setPowerProfileToNode(frequencies[i])
-		if err != nil {
-			fmt.Printf(fmt.Sprintf("ERR1 :%+v", err))
-		}
-
-		for j := 0; j < 1000; j++ {
-			url := "auth-python.default.192.168.1.240.sslip.io"
-			command := fmt.Sprintf("cd $HOME/vSwarm/tools/test-client && ./test-client --addr %s:80 --name \"allow\"", url)
-
-			startInvoke := time.Now().UTC().UnixMilli()
-			cmd := exec.Command("bash", "-c", command)
-			_, err := cmd.CombinedOutput()
-			if err != nil {
-				fmt.Printf(fmt.Sprintf("ERR2: %+v", err))
-				return
-			}
-			endInvoke := time.Now().UTC().UnixMilli()
-			latency := endInvoke - startInvoke
-			err = writer.Write(append([]string{strconv.FormatInt(startInvoke, 10), strconv.FormatInt(endInvoke, 10), strconv.FormatInt(latency, 10)}))
-			if err != nil {
-				fmt.Printf("Error writing metrics to the CSV file: %v\n", err)
-			}
-		}
-
-		err = writer.Write(append([]string{"-", "-", "-"}))
-		if err != nil {
-			fmt.Printf("Error writing metrics to the CSV file: %v\n", err)
-		}
-		fmt.Println("done")
-	}
-}
