+++ scripts/power_manager/powerconfig.yaml
