+++ scripts/power_manager/shared-workload.yaml
@@ -1,11 +1,11 @@
 apiVersion: "power.intel.com/v1"
 kind: PowerWorkload
 metadata:
-  name: shared-node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us-workload
+  name: performance-node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us-workload
   namespace: intel-power
 spec:
-  name: "shared-node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us-workload"
+  name: "performance-node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us-workload"
   allCores: true
   powerNodeSelector:
     kubernetes.io/hostname: node-1.kt-cluster.ntu-cloud-pg0.utah.cloudlab.us
-  powerProfile: "shared"
+  powerProfile: "performance"
